+++ distributedcombigrid/src/sgpp/distributedcombigrid/manager/ProcessGroupWorker.hpp
@@ -148,7 +148,7 @@ class ProcessGroupWorker {
 
   void processDuration(const Task& t, const Stats::Event e, unsigned int numProcs);
 
-  void updateTaskWithCurrentValues(Task& taskToUpdate, int numGrids);
+  void updateTaskWithCurrentValues(Task& taskToUpdate, size_t numGrids);
 };
 
 inline Task* ProcessGroupWorker::getCurrentTask() { return currentTask_; }
